New functionality to arc viz
