+++ superset/assets/src/visualizations/deckgl/layers/scatter.jsx
@@ -2,82 +2,30 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import PropTypes from 'prop-types';
 
 import { ScatterplotLayer } from 'deck.gl';
 
-import AnimatableDeckGLContainer from '../AnimatableDeckGLContainer';
-import Legend from '../../Legend';
-
+import CategoricalDeckGLContainer from '../CategoricalDeckGLContainer';
 import * as common from './common';
-import { getColorFromScheme, hexToRGB } from '../../../modules/colors';
-import { getPlaySliderParams } from '../../../modules/time';
 import { unitToRadius } from '../../../modules/geo';
-import sandboxedEval from '../../../modules/sandbox';
 
 
 function getPoints(data) {
   return data.map(d => d.position);
 }
 
-function getCategories(formData, payload) {
-  const fd = formData;
-  const c = fd.color_picker || { r: 0, g: 0, b: 0, a: 1 };
-  const fixedColor = [c.r, c.g, c.b, 255 * c.a];
-  const categories = {};
-
-  payload.data.features.forEach((d) => {
-    if (d.cat_color != null && !categories.hasOwnProperty(d.cat_color)) {
-      let color;
-      if (fd.dimension) {
-        color = hexToRGB(getColorFromScheme(d.cat_color, fd.color_scheme), c.a * 255);
-      } else {
-        color = fixedColor;
-      }
-      categories[d.cat_color] = { color, enabled: true };
-    }
-  });
-  return categories;
-}
-
-function getLayer(formData, payload, slice, filters) {
-  const fd = formData;
-  const c = fd.color_picker || { r: 0, g: 0, b: 0, a: 1 };
-  const fixedColor = [c.r, c.g, c.b, 255 * c.a];
-
-  let data = payload.data.features.map((d) => {
+function getLayer(fd, data, slice) {
+  const dataWithRadius = data.map((d) => {
     let radius = unitToRadius(fd.point_unit, d.radius) || 10;
     if (fd.multiplier) {
       radius *= fd.multiplier;
     }
-    let color;
-    if (fd.dimension) {
-      color = hexToRGB(getColorFromScheme(d.cat_color, fd.color_scheme), c.a * 255);
-    } else {
-      color = fixedColor;
-    }
-    return {
-      ...d,
-      radius,
-      color,
-    };
+    return { ...d, radius };
   });
-
-  if (fd.js_data_mutator) {
-    // Applying user defined data mutator if defined
-    const jsFnMutator = sandboxedEval(fd.js_data_mutator);
-    data = jsFnMutator(data);
-  }
-
-  if (filters != null) {
-    filters.forEach((f) => {
-      data = data.filter(f);
-    });
-  }
 
   return new ScatterplotLayer({
     id: `scatter-layer-${fd.slice_id}`,
-    data,
+    data: dataWithRadius,
     fp64: true,
     radiusMinPixels: fd.min_radius || null,
     radiusMaxPixels: fd.max_radius || null,
@@ -85,109 +33,6 @@ function getLayer(formData, payload, slice, filters) {
     ...common.commonLayerProps(fd, slice),
   });
 }
-
-const propTypes = {
-  slice: PropTypes.object.isRequired,
-  payload: PropTypes.object.isRequired,
-  setControlValue: PropTypes.func.isRequired,
-  viewport: PropTypes.object.isRequired,
-};
-
-class DeckGLScatter extends React.PureComponent {
-  /* eslint-disable-next-line react/sort-comp */
-  static getDerivedStateFromProps(nextProps) {
-    const fd = nextProps.slice.formData;
-
-    const timeGrain = fd.time_grain_sqla || fd.granularity || 'PT1M';
-    const timestamps = nextProps.payload.data.features.map(f => f.__timestamp);
-    const { start, end, step, values, disabled } = getPlaySliderParams(timestamps, timeGrain);
-
-    const categories = getCategories(fd, nextProps.payload);
-
-    return { start, end, step, values, disabled, categories };
-  }
-  constructor(props) {
-    super(props);
-    this.state = DeckGLScatter.getDerivedStateFromProps(props);
-
-    this.getLayers = this.getLayers.bind(this);
-    this.toggleCategory = this.toggleCategory.bind(this);
-    this.showSingleCategory = this.showSingleCategory.bind(this);
-  }
-  componentWillReceiveProps(nextProps) {
-    this.setState(DeckGLScatter.getDerivedStateFromProps(nextProps, this.state));
-  }
-  getLayers(values) {
-    const filters = [];
-
-    // time filter
-    if (values[0] === values[1] || values[1] === this.end) {
-      filters.push(d => d.__timestamp >= values[0] && d.__timestamp <= values[1]);
-    } else {
-      filters.push(d => d.__timestamp >= values[0] && d.__timestamp < values[1]);
-    }
-
-    // legend filter
-    if (this.props.slice.formData.dimension) {
-      filters.push(d => this.state.categories[d.cat_color].enabled);
-    }
-
-    const layer = getLayer(
-      this.props.slice.formData,
-      this.props.payload,
-      this.props.slice,
-      filters);
-
-    return [layer];
-  }
-  toggleCategory(category) {
-    const categoryState = this.state.categories[category];
-    categoryState.enabled = !categoryState.enabled;
-    const categories = { ...this.state.categories, [category]: categoryState };
-
-    // if all categories are disabled, enable all -- similar to nvd3
-    if (Object.values(categories).every(v => !v.enabled)) {
-      /* eslint-disable no-param-reassign */
-      Object.values(categories).forEach((v) => { v.enabled = true; });
-    }
-
-    this.setState({ categories });
-  }
-  showSingleCategory(category) {
-    const categories = { ...this.state.categories };
-    /* eslint-disable no-param-reassign */
-    Object.values(categories).forEach((v) => { v.enabled = false; });
-    categories[category].enabled = true;
-    this.setState({ categories });
-  }
-  render() {
-    return (
-      <div style={{ position: 'relative' }}>
-        <AnimatableDeckGLContainer
-          getLayers={this.getLayers}
-          start={this.state.start}
-          end={this.state.end}
-          step={this.state.step}
-          values={this.state.values}
-          disabled={this.state.disabled}
-          viewport={this.props.viewport}
-          mapboxApiAccessToken={this.props.payload.data.mapboxApiKey}
-          mapStyle={this.props.slice.formData.mapbox_style}
-          setControlValue={this.props.setControlValue}
-        >
-          <Legend
-            categories={this.state.categories}
-            toggleCategory={this.toggleCategory}
-            showSingleCategory={this.showSingleCategory}
-            position={this.props.slice.formData.legend_position}
-          />
-        </AnimatableDeckGLContainer>
-      </div>
-    );
-  }
-}
-
-DeckGLScatter.propTypes = propTypes;
 
 function deckScatter(slice, payload, setControlValue) {
   const fd = slice.formData;
@@ -202,11 +47,13 @@ function deckScatter(slice, payload, setControlValue) {
   }
 
   ReactDOM.render(
-    <DeckGLScatter
+    <CategoricalDeckGLContainer
       slice={slice}
-      payload={payload}
+      data={payload.data.features}
+      mapboxApiKey={payload.data.mapboxApiKey}
       setControlValue={setControlValue}
       viewport={viewport}
+      getLayer={getLayer}
     />,
     document.getElementById(slice.containerId),
   );
