+++ superset/viz.py
@@ -2355,6 +2355,12 @@ class DeckArc(BaseDeckGLViz):
     spatial_control_keys = ['start_spatial', 'end_spatial']
     is_timeseries = True
 
+    def query_obj(self):
+        fd = self.form_data
+        self.is_timeseries = bool(
+            fd.get('time_grain_sqla') or fd.get('granularity'))
+        return super(DeckArc, self).query_obj()
+
     def get_properties(self, d):
         dim = self.form_data.get('dimension')
         return {
